% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{align.curves}
\alias{align.curves}
\title{Align the curves
@param data Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@param k Do you love cats? Defaults to TRUE.
@keywords align.curves
@export
@examples
align.curves()}
\usage{
align.curves(data, days = 1:263, k = 3, n = 5)
}
\description{
Align the curves
@param data Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@param k Do you love cats? Defaults to TRUE.
@keywords align.curves
@export
@examples
align.curves()
}
