% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allsizes.R
\name{allsizes}
\alias{allsizes}
\title{Returns the list of objects according to 
increasing memory size}
\usage{
allsizes()
}
\value{
No value is returned
}
\description{
\code{allsizes()} returns the list of objects according to increasing memory size
}
\examples{
\dontrun{allsizes()}
}
