% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{cat_function}
\alias{cat_function}
\title{here's the line directly above the function. requires 'apos to see}
\usage{
cat_function(love = TRUE)
}
\description{
here's the line directly above the function. requires 'apos to see
}
