% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cluster.on.sums}
\alias{cluster.on.sums}
\title{Cluster mice based on cumulative sum of scores
@param c Do you love cats? Defaults to TRUE.
@keywords cluster.on.sums
@export
@examples
cluster.on.sums()}
\usage{
cluster.on.sums(c)
}
\description{
Cluster mice based on cumulative sum of scores
@param c Do you love cats? Defaults to TRUE.
@keywords cluster.on.sums
@export
@examples
cluster.on.sums()
}
