% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practices.R
\name{col_means}
\alias{col_means}
\title{Column Mean Function}
\usage{
col_means(df)
}
\arguments{
\item{df}{}
}
\value{
The Mean of a column
}
\description{
Column Mean Function
}
\details{
After inputting a dataframe it will output the mean of each column
}
\examples{
## Data_Frame <- data.frame (column1 = c(100, 150, 120), column2 = c(60, 30, 45))
## col_means(Data_Frame)
}
\author{
Andrew Hwang
}
