% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare.auc}
\alias{compare.auc}
\title{Compare AUCs
@param data Do you love cats? Defaults to TRUE.
@param meta Do you love cats? Defaults to TRUE.
@param by Do you love cats? Defaults to TRUE.
@param test Do you love cats? Defaults to TRUE.
@keywords compare.auc
@export
@examples
compare.auc()}
\usage{
compare.auc(dataf, meta, by = "sex", test = "t.test")
}
\description{
Compare AUCs
@param data Do you love cats? Defaults to TRUE.
@param meta Do you love cats? Defaults to TRUE.
@param by Do you love cats? Defaults to TRUE.
@param test Do you love cats? Defaults to TRUE.
@keywords compare.auc
@export
@examples
compare.auc()
}
