% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practices.R
\name{count_na}
\alias{count_na}
\title{Count NA Function}
\usage{
count_na(vec)
}
\arguments{
\item{vec}{A numeric or character vector.}
}
\value{
An integer representing the count of NA values in the vector.
}
\description{
This function counts the number of NA values in a given vector using a for-loop.
}
\details{
After inputting a vector, it will return the count of NA values.
}
\examples{
vec <- c(1, 2, NA, 4, NA, 6)
count_na(vec)

}
\author{
Andrew Hwang
}
