% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cumsum}
\alias{cumsum}
\title{Compute cumulative sums
@param data Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@keywords cumsum
@export
@examples
cumsum()}
\usage{
cumsum(data, days = 1:263)
}
\description{
Compute cumulative sums
@param data Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@keywords cumsum
@export
@examples
cumsum()
}
