\docType{data}
\name{dataMLANP}
\alias{antropometric}
\alias{gaussian1_128}
\alias{exercise.trial1}
\alias{exercise.trial2}
\alias{students.survey}
\alias{students.test}
\alias{exercise.antr1}
\alias{children1}
\alias{children.rid}
\alias{firms}
\alias{soccer1}
\alias{exercise_child1}
\alias{exercise_soccer1}
\alias{exercise1}
\alias{exercise2}
\alias{granprix}
\title{Data sets of MLANP used in my lessons and exercises}
\description{
Data sets used in my lessons and labs. }
\usage{"data('name_of_data')"}
\format{use str("name_of_data")}
\examples{
data(antropometric)
str(antropometric)
data(exercise.trial1)
str(exercise.trial1)
data(exercise.trial2)
str(exercise.trial2)
data(students.survey)
str(students.survey)
data(students.test)
str(students.test)
data(exercise.antr1)
str(exercise.antr1)
data(children1)
str(children1)
data(children.rid)
str(children.rid)
data(firms)
str(firms)
data(soccer1)
str(soccer1)
data(exercise_child1)
str(exercise_child1)
data(exercise_soccer1)
str(exercise_soccer1)
data(exercise1)
str(exercise1)
data(exercise2)
str(exercise2)
data(granprix)
str(granprix)
}
\keyword{datasets}
\keyword{antropometric}
\keyword{children}
\keyword{trials}
\keyword{students}
\keyword{granprix}
