% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deck_function.R
\name{deck_function}
\alias{deck_function}
\title{A deck function based on team member status}
\usage{
deck_function(team_member = TRUE)
}
\arguments{
\item{team_member}{Are you part of the deck team? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
deck_function()
}
\keyword{deck}
