% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawbivnorm.R
\name{drawbivnorm}
\alias{drawbivnorm}
\title{drawbivnorm}
\usage{
drawbivnorm(rho = 0, sigma2 = 1)
}
\arguments{
\item{rho}{correlation between components}

\item{sigma2}{standard deviation of 
the second variable (the first is fixed to one)}
}
\value{
An interactive 3d plot
}
\description{
Interactive visualization ov bivariate normal density
}
\details{
Interactive didactic example of bivariate normal density
(centered on the origin)
trough an interactive panel which controls a 3d plot 
allowing \code{rho} and \code{sigma2} to vary through sliders. 
The other scale parameter is fixed to one.
}
\examples{
\dontrun{drawbivnorm()}	


}
