% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlarged.range.R
\name{enlarged.range}
\alias{enlarged.range}
\title{Enlarge the range of a variable}
\usage{
enlarged.range(x, eps = 0.05)
}
\arguments{
\item{x}{a numerical variable}

\item{eps}{increment of the range}
}
\value{
the range of x enlarged of a proportion of \code{eps/2} on each side
}
\description{
Enlarge the range of a variable
}
