% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahr_to_kelvin.R
\name{fahr_to_kelvin}
\alias{fahr_to_kelvin}
\title{Convert Fahrenheit to Kelvin}
\usage{
fahr_to_kelvin(temp)
}
\arguments{
\item{temp}{The input temperature.}
}
\description{
This function converts input temperatures in Fahrenheit to Kelvin.
}
\examples{
fahr_to_kelvin(32)
}

