% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find.auc}
\alias{find.auc}
\title{Compute AUC
@param data Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@param min.score Do you love cats? Defaults to TRUE.
@keywords find.auc
@export
@examples
find.auc()}
\usage{
find.auc(data, days = 1:263, min.score = 0)
}
\description{
Compute AUC
@param data Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@param min.score Do you love cats? Defaults to TRUE.
@keywords find.auc
@export
@examples
find.auc()
}
