\name{italycatalog}
\docType{data}
\alias{italycatalog}
\title{Small sample catalog of italian earthquakes}
\description{
A small sample catalog of italian earthquakes of magnitude at least 3.0 from year 2005 to year 2013.}
\usage{italycatalog}
\format{a data matrix with 2,158 observations and 5 variables: \code{time}, \code{lat}, \code{long}, \code{z}, \code{magn1}.
}
\examples{
data(italycatalog)
str(italycatalog)
}
\source{INGV (Istituto Nazionale di Geofisica e Vulcanologia) ISIDE Data Base.}
\references{INGV  home page: \url{http://www.ingv.it}.}
\keyword{datasets}
\keyword{earthquake}
