% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mahala.factor.R
\name{Mahala.factor}
\alias{Mahala.factor}
\title{Mahala.factor}
\usage{
Mahala.factor(x, eps = 1e-07, plot = FALSE)
}
\arguments{
\item{x}{a matrix with observed data}

\item{eps}{minimum scaled value of eigenvalues}

\item{plot}{logical. If \code{plot=TRUE} a qq plot of distances vs. 
theoretical chi-square distribution is plotted}
}
\value{
The vector of Mahalanobis disctances from centroid
}
\description{
Computes the Mahalanobis distances of dat points from their centroids.
Uses the spectral decomposition of Variance matrix for numerical stability
}
