% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahala.x.R
\name{mahala.x}
\alias{mahala.x}
\title{Computes the Mahalanobis distances of data points from  a point.}
\usage{
mahala.x(x, mu = apply(x, 2, mean), sigma = cov(x), qplot = FALSE)
}
\arguments{
\item{x}{input data matrix}

\item{mu}{vector of center from which distances are computed}

\item{sigma}{given covariance matrix}

\item{qplot}{logical. if TRUE plots the final qq-plot of distances}
}
\value{
Mahalanobis distances
}
\description{
Computes the Mahalanobis distances of data points
 from  a given point or from the centroid. 
 The covariance matrix  can be given as input or computed on data
}
\details{
Not computationally efficient. Someday it will be
joined with Mahala.factor. I do not remember why I wrote 
two different routines...
}
