% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLA.bivariate.normal.R
\name{MLA.bivariate.normal}
\alias{MLA.bivariate.normal}
\title{Plts of 2d  contour ellipse of a bivariate normal distribution,
 optionally with random points}
\usage{
MLA.bivariate.normal(
  rho = 0.5,
  mu = c(0, 0),
  s = c(1, 1),
  lev = c(0.25, 0.5, 0.75),
  vec.col = 1:(length(lev) + 1),
  axes = FALSE,
  npoints = 0,
  ...
)
}
\arguments{
\item{rho}{a scalar: correlation of components}

\item{mu}{a two-elements vector of means}

\item{s}{a two-elements vector of standard deviation}

\item{lev}{a vector of levels for the 2d contour ellipses}

\item{vec.col}{a vector of colurs numbers for the 2d contour ellipses}

\item{axes}{if \code{TRUE} principal axes will be drawn 
(no axes drawn at the present moment)}

\item{npoints}{number of points drawn 
from the bivariate normal distribution and 
coloured according to vec.col}

\item{...}{graphical parameters for points drawing}
}
\value{
only plotting is done
}
\description{
\code{MLA.bivariate.normal} plots a bivariate normal distribution
with 2d-ellipse contours at different levels,
 optionally with random points
}
\details{
The complete representations draws ellipses together 
with points, coloured according to the level of density. Also principal
axes can be drawn.
}
