% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLA.contour3d.R
\name{MLA.contour3d}
\alias{MLA.contour3d}
\title{3d double contour plot of a  kernel density estimator compared 
with normal contour ellipses}
\usage{
MLA.contour3d(
  data.obs,
  ind.var = 1:3,
  standardize = TRUE,
  animation = FALSE,
  n.rounds = 1,
  princ.axes = TRUE,
  color.points = 1,
  xy = TRUE,
  xz = TRUE,
  yz = TRUE,
  marg.col = "lightcyan",
  ellipse = FALSE,
  levels = c(0.25, 0.5, 0.75)
)
}
\arguments{
\item{data.obs}{A matrix of observed data set with least 3 columns.}

\item{ind.var}{Vector of integers (3 elements): which variables must be used?}

\item{standardize}{If \code{TRUE}  variables will be standardized before PCA}

\item{animation}{If \code{animation=TRUE} the final visualization will be animated}

\item{n.rounds}{number of animated rotations (10 seconds each)}

\item{princ.axes}{If \code{princ.axes=TRUE} principal axes will be drawn}

\item{color.points}{color for observed points}

\item{xy}{If \code{xy=TRUE} projection on a xy plan will be drawn}

\item{xz}{If \code{xz=TRUE} projection on a xz plan will be drawn}

\item{yz}{If \code{yz=TRUE} projection on a yz plan will be drawn}

\item{marg.col}{color for marginal projection}

\item{ellipse}{If \code{ellipse=TRUE} 3d contour ellipses will be drawn at probability given by \code{levels}}

\item{levels}{probability levels a which 3d contour ellipses will be drawn}
}
\value{
A 3d interactive double plot
}
\description{
3d contour plot of a  kernel density estimator of a 3d data set, 
compared with normal contour ellipses
}
\examples{
\donttest{
data(antropometric)
MLA.contour3d(antropometric[,7:9],
standardize=FALSE,levels=c(.2,.4,.6,.8),
princ.axes=TRUE,ellipse=TRUE,animation=FALSE)
}

}
