% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLA.group.density.plot.R
\name{MLA.group.density.plot}
\alias{MLA.group.density.plot}
\title{Density plot of a grouped variable}
\usage{
MLA.group.density.plot(x, ind, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{ind}{a grouping variable (same length of \code{x})}

\item{...}{plotting options}
}
\value{
A plot with the density of  \code{x} grouped by  \code{ind}
}
\description{
Density plot of a grouped variable
}
\examples{
\donttest{
data(children.rid)
MLA.explor.plot2D(children.rid$gestazione,children.rid$lunghezza)
MLA.group.density.plot(children.rid$lunghezza,children.rid$gestazione)
}
}
