% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLA.pca3d.R
\name{MLA.pca3d}
\alias{MLA.pca3d}
\title{MLA.pca3d}
\usage{
MLA.pca3d(
  data.obs,
  ind.var = 1:3,
  ind.comp = 1:3,
  standardize = TRUE,
  lim.max = 4,
  segments = FALSE,
  animation = FALSE,
  princ.axes = TRUE,
  color.points = 1,
  xy = TRUE,
  xz = TRUE,
  yz = TRUE,
  marg.col = "lightcyan",
  ellipse = TRUE,
  levels = c(0.25, 0.5, 0.75),
  lim.stand = TRUE,
  only3vars = TRUE
)
}
\arguments{
\item{data.obs}{A matrix of observed data set with least 3 columns.}

\item{ind.var}{Vector of integers (at least 3 elements): which variables must be used?}

\item{ind.comp}{Vector of integers ( 3 elements): which components must be displayed?}

\item{standardize}{If \code{TRUE} (strongly suggested) variables will be standardized before PCA}

\item{lim.max}{Maximum absolute standardized limit}

\item{segments}{If \code{segments=TRUE} Segments starting from each point, 
orthogonal to the first principal axis, will be drawn}

\item{animation}{If \code{animation=TRUE} the final visualization will be animated}

\item{princ.axes}{If \code{princ.axes=TRUE} principal axes will be drawn}

\item{color.points}{Color for points}

\item{xy}{If \code{xy=TRUE} projection on a xy plan will be drawn}

\item{xz}{If \code{xz=TRUE} projection on a xz plan will be drawn}

\item{yz}{If \code{yz=TRUE} projection on a yz plan will be drawn}

\item{marg.col}{color for marginal projection}

\item{ellipse}{If \code{ellipse=TRUE} 3d contour ellipses will be drawn at probability given by \code{levels}}

\item{levels}{probability levels a which 3d contour ellipses will be drawn}

\item{lim.stand}{If \code{lim.stand=TRUE} fixed standardized limits 
of  \code{lim.max=TRUE} will be used for x,y,z}

\item{only3vars}{use only three variables}
}
\value{
3d plots with many options
}
\description{
3d Principal Components Analysis
}
\examples{
\donttest{
data(antropometric)
MLA.pca3d(antropometric[,7:13])
}
}
