% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLA.plot.norm.R
\name{MLA.plot.norm}
\alias{MLA.plot.norm}
\title{Histogram, non parametric density, normal density}
\usage{
MLA.plot.norm(z, nc = 50, std = FALSE)
}
\arguments{
\item{z}{input numerical variable}

\item{nc}{number of classes for histogram}

\item{std}{if \code{std=TRUE} then variable 
is standardized before plotting}
}
\value{
plots the histogram of the distribution of the input variable
together with a non parametric estimate of the density and
  a normal density.
}
\description{
Join to MLA.plot.feq some day...
}
\details{
The function plots the histogram of the distribution of the input variable
together with a non parametric estimate of the density and
  a normal density. Possibly \code{x} is a vector of 
  \code{lm$residuals}.
}
\examples{
MLA.plot.norm(rnorm(100))
MLA.plot.norm(rgamma(1000,shape=2))
}
