% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLA.quantile.plot.R
\name{MLA.quantile.plot}
\alias{MLA.quantile.plot}
\title{Plot quantile lines}
\usage{
MLA.quantile.plot(y, x, seg = TRUE, probs = seq(0, 1, 0.25))
}
\arguments{
\item{y}{a numerical variable}

\item{x}{a grouping variable}

\item{seg}{logical. If \code{seg=TRUE }vertical segments are drawn}

\item{probs}{vector of probabilities for which quantiles are computed.
Default values \code{seq(0, 1, 0.25) }}
}
\value{
A plot with quantiles plot
}
\description{
Plot quantile lines
}
\examples{
\donttest{
data(children.rid)
MLA.explor.plot2D(children.rid$gestazione,children.rid$lunghezza)
MLA.quantile.plot(children.rid$lunghezza,children.rid$gestazione)
}
}
