% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLA.regression3d.R
\name{MLA.regression3d}
\alias{MLA.regression3d}
\title{Visualization of a multiple linear regression with two regressors}
\usage{
MLA.regression3d(
  data.obs,
  ext = 0.3,
  newplot = TRUE,
  points = TRUE,
  conf.exp = TRUE,
  conf.sing = FALSE,
  level = 0.95,
  ell = FALSE,
  liv.ell = 0.2,
  ngrid = 50,
  rotation = FALSE,
  col.obs = 1,
  col.fit = 2
)
}
\arguments{
\item{data.obs}{A matrix of observed data set with 3 columns, \code{x1,x2,y}}

\item{ext}{proportion of extension of ranges of \code{x1,x2}}

\item{newplot}{logical. if \code{newplot=TRUE} a new 3d device is opened}

\item{points}{logical; if \code{points=TRUE} points are plotted}

\item{conf.exp}{logical; if \code{ conf.exp=TRUE} confidence surfaces for expected values are drawn}

\item{conf.sing}{logical; if \code{conf.sing=TRUE} confidence surfaces for future  observations are drawn}

\item{level}{numerical; confidence level for confidence surfaces}

\item{ell}{logical; if \code{ ell=TRUE}  ellipses for the conjoint  distribution of \code{x1,x2} are drawn at a prob. level given by \code{liv.ell}.}

\item{liv.ell}{numerical; prob. level for the \code{x1,x2}ellipses}

\item{ngrid}{integer. Number of grid points for each axis  in \code{x1,x2} plane to compute confidence surfaces}

\item{rotation}{a rotation of the final 3d view is performed}

\item{col.obs}{color of observed points}

\item{col.fit}{color of fitted points}
}
\value{
no value is returned: only 3d plotting is performed
}
\description{
Visualization of a multiple linear regression with confidence
region for expected values   \code{E(y)=beta0+beta1*x1+beta2*x2} 
and for future observations  input is a matrix with \code{x1,x2,y}
}
\examples{
\donttest{
data("antropometric")
MLA.regression3d(antropometric[101:300,c(7,9,8)],conf.exp=FALSE,level=0.99)
open3d()
MLA.regression3d(antropometric[101:200,c(7,9,8)],level=0.99)
open3d()
MLA.regression3d(antropometric[101:200,c(7,13,8)],level=0.99,ell=TRUE)}



}
