% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLA.summ.mat.R
\name{MLA.summ.mat}
\alias{MLA.summ.mat}
\title{MLA.summ.mat}
\usage{
MLA.summ.mat(data)
}
\arguments{
\item{data}{a matrix or a data.frame}
}
\value{
a matrix with \code{nrow(data)}: 
        each row contains a summary output given by \code{summ}
}
\description{
a \emph{personal} summary of the variables of a matrix or a data.frame
\emph{Not computationally optimized}
}
\examples{
\donttest{
data(antropometric)
print(round(MLA.summ.mat(antropometric[,7:13]),2))}
}
