% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLA.trivariate.normal.R
\name{MLA.trivariate.normal}
\alias{MLA.trivariate.normal}
\title{Plots a trivariate normal distribution
with 3d-ellipse contours,
 optionally with random points}
\usage{
MLA.trivariate.normal(
  sigma = matrix(c(1, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 0.5, 1), 3, 3),
  mu = c(0, 0, 0),
  lev = c(0.25, 0.5, 0.75),
  vec.col = 1:(length(lev) + 1),
  grid.ell = TRUE,
  axes = TRUE,
  ext = 4,
  zplane.cond = TRUE,
  z.val = 1,
  npoints = 0,
  ...
)
}
\arguments{
\item{sigma}{a 3x3 variance-covariance matrix}

\item{mu}{a vector of three mean values}

\item{lev}{a vector of levels for the 3d contour ellipses}

\item{vec.col}{a vector of colurs numbers for the 3d contour ellipses}

\item{grid.ell}{if TRUE a mesh is plotted on the ellipses surfaces}

\item{axes}{if TRUE principal axes of the ellipses are drawn}

\item{ext}{extension, in standardized units, of principal axes}

\item{zplane.cond}{if TRUE a horizontal plane is plotted at level z.val}

\item{z.val}{value at which a horizontal plane is plotted}

\item{npoints}{number of points drawn 
from the trivariate normal distribution and 
coloured according to vec.col}

\item{...}{graphical parameters for points drawing}
}
\value{
a 3d plot of a trivariate normal. According to the options used, 
a sample is drawn and plotted using the same colour used for the nearest external ellipse. 
(points inside a most internal black ellipse will be black )
}
\description{
\code{MLA.trivariate.normal} plots a trivariate normal distribution
with 3d-ellipse contours at different levels,
 optionally with random points
}
\details{
The complete representations draws ellipses together 
with points, coloured according to the level of density. Also principal
axes can be drawn.
}
\examples{
\donttest{
MLA.trivariate.normal(npoints=1000)
MLA.trivariate.normal(sigma=matrix(c(1,0.0,0.0,0.0,1,0.8,0.0,0.8,1),3,3),npoints=100000,size=1)
MLA.trivariate.normal(sigma=matrix(c(1,0.0,0.0,0.0,1,0.8,0.0,0.8,1),3,3),lev=0.5, npoints=1000)
}
}
