\name{MLANP-internal}
\title{Internal MLANP functions}
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\alias{likelygen}

\description{Internal functions used in the package MLANP }
\author{Marcello Chiodi}
\details{These are not intended to be used directly by users}
\keyword{internal}