\name{MLANP-package}
\alias{MLANP-package}
\docType{package}
\title{Miscellaneous functions by Marcello Chiodi}
\description{Simple functions for my lectures on linear models and on non parametric statical methods}
\details{
\tabular{ll}{
Package: MLANP
Type: \tab Package\cr
Version: \tab 1.4.1\cr
Date: \tab 2020-04-19 \cr
License: \tab GPL2 \cr
}
}
\author{Marcello Chiodi}
\examples{
\donttest{

example(rnormal.mix2biv)
example(MLA.explor.pairs)
example(MLA.explor.plot2D)
example(MLA.summ.mat)

example(MLA.pca3d)
example(MLA.trivariate.normal)
example(MLA.contour3d)

example(simul.normalmix2)


}
}