% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.grouped.bin.R
\name{myplot.grouped.bin}
\alias{myplot.grouped.bin}
\title{Plot the means of a zero-one response variable as a function
 of a numerical grouping variable}
\usage{
myplot.grouped.bin(
  y,
  group,
  rid = 5,
  ksd = 2,
  lines0 = TRUE,
  interval = TRUE,
  ylim0 = range(y),
  fac.range = 0.5,
  line.color = 1,
  back.color = 8,
  ...
)
}
\arguments{
\item{y}{a zero-one response variable (or however a 
variable with only two distinct values)}

\item{group}{a numerical grpouping variable}

\item{rid}{a factor to adjust the size of the circles}

\item{ksd}{multiple of the standard deviation to plot vertical segments}

\item{lines0}{if \code{lines0=TRUE} a line connecting means is drawn}

\item{interval}{if \code{interval=TRUE} plot vertical segments are drawn}

\item{ylim0}{y range of the plot}

\item{fac.range}{factor of enlargement for y-axis range}

\item{line.color}{color of the lines}

\item{back.color}{background color}

\item{...}{other graphical parameters}
}
\value{
a list with \code{y} means, grouping values, sizes
}
\description{
Plot the means of a zero-one response variable as a function
 of a numerical grouping variable
}
\examples{
\donttest{
data(students.test)
myplot.grouped.bin(students.test$imm_test,students.test$VOTO_B_100)}
}
