% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages-package.R
\docType{package}
\name{packages-package}
\alias{packages}
\alias{packages-package}
\title{packages: STAT 627}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Xubo Tang \email{xt0610a@american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
