% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{packages}
\alias{packages}
\title{Package Loading}
\usage{
packages(packages)
}
\arguments{
\item{packages}{A string or string vector with the names of the packages to load..}
}
\description{
This function loads specified R packages, after having made sure they are installed.
}
\examples{
packages("lme4")
}
\keyword{packages}

