% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.H1.R
\name{panel.H1}
\alias{panel.H1}
\title{Panel plot called by MLA.explor.pairs}
\usage{
panel.H1(x, ...)
}
\arguments{
\item{x}{input data}

\item{...}{other parameters to pass}
}
\value{
a panel plot
}
\description{
Panel plot called by MLA.explor.pairs
}
