% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.lmsm.R
\name{panel.lmsm}
\alias{panel.lmsm}
\title{Panel plot called by MLA.explor.pairs}
\usage{
panel.lmsm(
  x,
  y,
  col = par("col"),
  bg = NA,
  pch = par("pch"),
  cex = 1,
  col.smooth = "red",
  span = 2/3,
  iter = 3,
  ...
)
}
\arguments{
\item{x}{input data}

\item{y}{input data}

\item{col}{color of points}

\item{bg}{background color of points}

\item{pch}{graphical par}

\item{cex}{graphical par}

\item{col.smooth}{color of smoothing line}

\item{span}{smoothing}

\item{iter}{number of iterations in smoothing}

\item{...}{other parameters to pass}
}
\value{
a panel plot
}
\description{
Panel plot called by MLA.explor.pairs
}
