% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.line3d.R
\name{parametric.line3d}
\alias{parametric.line3d}
\title{Gives the three directions of the intersection line of two planes}
\usage{
parametric.line3d(plane1, plane2)
}
\arguments{
\item{plane1}{a vector of three parameters \code{a,b,c} of a plane of equation
\code{a x + b y + c z = k}}

\item{plane2}{plane1 a vector of three parameters \code{a,b,c} of a plane of equation
\code{a x + b y + c z = k}}
}
\value{
A vector of the three direction of the intersection line
}
\description{
Gives the three directions of the intersection line of two planes
}
