\name{pareto-package}
\alias{pareto-package}
\alias{pareto}
\docType{package}
\title{
\packageTitle{pareto}
}
\description{
\packageDescription{pareto}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pareto}
\packageIndices{pareto}
density function for pareto distribution
}
\author{
\packageAuthor{pareto}

Maintainer: \packageMaintainer{pareto}
}

\keyword{ package }
