% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.curve}
\alias{plot.curve}
\title{Plot the mean curves for the groups
@param dataf Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@param show Do you love cats? Defaults to TRUE.
@param min.score Do you love cats? Defaults to TRUE.
@param by Do you love cats? Defaults to TRUE.
@keywords plot.curve
@export
@examples
plot.curve()}
\usage{
\method{plot}{curve}(dataf, days = 1:263, show = "None", min.score = 0, by = "sex")
}
\description{
Plot the mean curves for the groups
@param dataf Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@param show Do you love cats? Defaults to TRUE.
@param min.score Do you love cats? Defaults to TRUE.
@param by Do you love cats? Defaults to TRUE.
@keywords plot.curve
@export
@examples
plot.curve()
}
