% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.curves}
\alias{plot.curves}
\title{Plot the curves for the groups by mice
@param data Do you love cats? Defaults to TRUE.
@param group Do you love cats? Defaults to TRUE.
@param min.score Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@keywords plot.curves
@export
@examples
plot.curves()}
\usage{
\method{plot}{curves}(data, group, min.score = 0, days = 1:263)
}
\description{
Plot the curves for the groups by mice
@param data Do you love cats? Defaults to TRUE.
@param group Do you love cats? Defaults to TRUE.
@param min.score Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@keywords plot.curves
@export
@examples
plot.curves()
}
