% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.mean}
\alias{plot.mean}
\title{Plot the means along a period of time
@param data Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@param min.score Do you love cats? Defaults to TRUE.
@keywords plot.mean
@export
@examples
plot.mean()}
\usage{
\method{plot}{mean}(data, days = 1:263, min.score = 0)
}
\description{
Plot the means along a period of time
@param data Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@param min.score Do you love cats? Defaults to TRUE.
@keywords plot.mean
@export
@examples
plot.mean()
}
