% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbiasv.R
\name{plotbiasv}
\alias{plotbiasv}
\title{Plot of comparison of bias vs variance (lesson exercise)}
\usage{
plotbiasv(x, ...)
}
\arguments{
\item{x}{output of simul.normalmix2}

\item{...}{other arguments to pass}
}
\value{
a plot of comparison of bias vs variance
}
\description{
Plot of comparison of bias vs variance (lesson exercise)
}
\examples{
\donttest{
d1=simul.normalmix2(n=1000,ncamp=500,ngrid=200,tpause = 0)
plotbiasv(d1)}
}
