% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnormal.mix2biv.R
\name{rnormal.mix2biv}
\alias{rnormal.mix2biv}
\title{Generates bivariate vectors from a mixture
 of two bivariate normal 
 distributions}
\usage{
rnormal.mix2biv(
  n = 100,
  mean = rbind(c(0, 0), c(2, 2)),
  sd = rbind(c(1, 1), c(1, 1)),
  r = c(0, 0),
  p = 0.5
)
}
\arguments{
\item{n}{number of bivariate vectors to generate}

\item{mean}{two vectors of means (one for each component)}

\item{sd}{two vectors of standard deviations (one for each component)}

\item{r}{two values of correlation (one for each component)}

\item{p}{weight parameter}
}
\value{
a nx2 matrix of random bivariate vectors
}
\description{
Generates bivariate vectors from a mixture
 of two bivariate normal 
 distributions with mixing proportion p and (1-p)
}
\examples{
\donttest{
plot(rnormal.mix2biv(10000,r=c(0,-0.8),p=0.2),pch=".")}

}
