% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sickMice.proportion}
\alias{sickMice.proportion}
\title{Return proportion of sick mice per day
@param data Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@keywords sickMice.proportion
@export
@examples
sickMice.proportion()}
\usage{
sickMice.proportion(data, days = 1:263)
}
\description{
Return proportion of sick mice per day
@param data Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@keywords sickMice.proportion
@export
@examples
sickMice.proportion()
}
