% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stat.comp}
\alias{stat.comp}
\title{Compare the mean scores between 2 days
@param data Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@param min.score Do you love cats? Defaults to TRUE.
@param test Do you love cats? Defaults to TRUE.
@keywords stat.comp
@export
@examples
stat.comp()}
\usage{
stat.comp(data, days = c(1, 263), min.score = 0, test = "t.test")
}
\description{
Compare the mean scores between 2 days
@param data Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@param min.score Do you love cats? Defaults to TRUE.
@param test Do you love cats? Defaults to TRUE.
@keywords stat.comp
@export
@examples
stat.comp()
}
