% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLA.summ.mat.R
\name{summ}
\alias{summ}
\title{A \emph{personal} summary of a numerical variable}
\usage{
summ(x)
}
\arguments{
\item{x}{a numerical array}
}
\value{
A list of summary statistics
\itemize{
\item n number of observations
\item minimum, maximum
\item mean, mean square deviation
\item median, mean absolute deviatons from median and from mean
\item gamma1,gamma2: skewness and kurtosis indeces (0 for the normal,gamma2=beta2-3)
}
}
\description{
a \emph{personal} summary of a numerical variable. 
\emph{Not computationally optimized}
}
\examples{
\donttest{summ(rnorm(1000))}
}
